feat: No Dots on a Single Card
